Extract assertion messages to doctest/parts/assert/message.h
